Move account-create sub-components together.
