+++ assets/js/modules/analytics/common/account-create.js
@@ -28,11 +28,11 @@ import { useCallback, useState, useEffect } from '@wordpress/element';
 import Button from '../../../components/button';
 import ProgressBar from '../../../components/progress-bar';
 import { trackEvent } from '../../../util';
-import TimezoneSelect from './timzezone-select';
-import AccountField from './account-field';
-import PropertyField from './property-field';
-import ProfileField from './profile-field';
-import CountrySelect from './country-select';
+import TimezoneSelect from './account-create/timezone-select';
+import AccountField from './account-create/account-field';
+import PropertyField from './account-create/property-field';
+import ProfileField from './account-create/profile-field';
+import CountrySelect from './account-create/country-select';
 import { STORE_NAME, FORM_ACCOUNT_CREATE } from '../datastore/constants';
 import { STORE_NAME as CORE_SITE } from '../../../googlesitekit/datastore/site/constants';
 import { countryCodesByTimezone } from '../util/countries-timezones';
