+++ assets/js/modules/analytics/common/account-create/profile-field.js
@@ -1,5 +1,5 @@
 /**
- * AccountField component.
+ * ProfileField component.
  *
  * Site Kit by Google, Copyright 2020 Google LLC
  *
@@ -25,27 +25,27 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import CreateAccountField from './create-account-field';
+import CreateAccountField from '../create-account-field';
 import Data from 'googlesitekit-data';
-import { STORE_NAME, FORM_ACCOUNT_CREATE } from '../datastore/constants';
+import { STORE_NAME, FORM_ACCOUNT_CREATE } from '../../datastore/constants';
 
 const { useSelect, useDispatch } = Data;
 
-export default function AccountField() {
-	const value = useSelect( ( select ) => select( STORE_NAME ).getForm( FORM_ACCOUNT_CREATE, 'accountName' ) );
-
+export default function ProfileField() {
+	const value = useSelect( ( select ) => select( STORE_NAME ).getForm( FORM_ACCOUNT_CREATE, 'profileName' ) );
 	const { setForm } = useDispatch( STORE_NAME );
-	const setValue = useCallback( ( accountName ) => {
-		setForm( FORM_ACCOUNT_CREATE, { accountName } );
+
+	const setValue = useCallback( ( profileName ) => {
+		setForm( FORM_ACCOUNT_CREATE, { profileName } );
 	}, [ setForm ] );
 
 	return (
 		<CreateAccountField
-			label={ __( 'Account', 'google-site-kit' ) }
-			hasError={ ! value }
+			label={ __( 'View', 'google-site-kit' ) } // Profiles are now called views.
 			value={ value }
+			hasError={ ! value }
 			setValue={ setValue }
-			name="account"
+			name="profile"
 		/>
 	);
 }
